Add addbook.js module to handle book adding
